Add not ready option -a
Add option -a to analyze file that we can recive by redirect include-explorer output to file.
The result of this option is list of include files from some target directory.
For example (+: valid, -:invalid)

+ .    target_project/main.h - project file from project file
+ ..   target_project/foo.h - project file from project file
+ ...  /usr/include/stdio.h - system file from project file
- .... /usr/include/wchar.h - system file from system file
+ ..   /usr/include/stdio.h - system file from project file

Using:
./include-explorer -f test/compile_commands.json > res.txt 2>&1
./include-explorer -a res.txt include-explorer/test

Now it does not work properly. Now it only show include files with path and level.